+++ CHANGELOG.md
@@ -11,6 +11,7 @@ levelAcceleration - defined by map car spawn delay (in ticks, on the start of th
 
 * planned: [fix] that the power remains equipped after game is reset but cannot be used
 ## live
+* [fix] Fixed background bug on GPU that not supports non power of two sizes of textures.
 * [fix] Added check for resource folder (game is not crashing when error)
 * [technical] Added some command-line arguments
 ## v200302